Move architecture independent io.c/io.h out of x86
